Fix build issues and enhance project structure: Remove turbopack, add TypeScript error handling, refactor components
