+++ .eslintrc.js
@@ -2,11 +2,3 @@ module.exports = {
   extends: [
     'next/core-web-vitals',
     'plugin:@typescript-eslint/recommended'
-  ],
-  rules: {
-    '@typescript-eslint/no-explicit-any': 'error',
-    '@typescript-eslint/no-unused-vars': 'error',
-    '@typescript-eslint/no-require-imports': 'error',
-    'react-hooks/exhaustive-deps': 'warn'
-  }
-};
